AMP-10498
git-svn-id: file:///svn/branches/AMP_1_16_RELEASE/amp@16060 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/helper/ActivityItem.java
@@ -72,7 +72,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
 
 	private ActivityUtil.ActivityAmounts amounts;
     // percent of specific program assigned
-    private Long percent;
+    private Float percent;
     // used only for program percent applying
     private AmpActivity act;
 
@@ -101,7 +101,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
      * @see ActivityItem#ActivityItem(AmpActivity)
 	 */
 
-	public ActivityItem(AmpActivity act,Long percent) {
+	public ActivityItem(AmpActivity act,Float percent) {
         this.act=act;
         this.percent=percent;
 
@@ -117,7 +117,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
 		this(entity,new SimpleDateFormat(Constants.CALENDAR_DATE_FORMAT),"USD",null, request);
 	}
 
-    public ActivityItem(AmpActivity entity,String curenncyCode,Long percent, HttpServletRequest request) throws Exception{
+    public ActivityItem(AmpActivity entity,String curenncyCode,Float percent, HttpServletRequest request) throws Exception{
         this(entity,new SimpleDateFormat(Constants.CALENDAR_DATE_FORMAT),curenncyCode,percent, request);
 	}
 
@@ -127,7 +127,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
 	 * @param entity AmpActivity db entity to construct helper from
 	 * @param frmt date formatter
 	 */
-	public ActivityItem(AmpActivity entity,DateFormat frmt,String curenncyCode,Long percent, HttpServletRequest request) throws Exception {
+	public ActivityItem(AmpActivity entity,DateFormat frmt,String curenncyCode,Float percent, HttpServletRequest request) throws Exception {
 		Site site = RequestUtils.getSite(request);
 		Locale navigationLanguage = RequestUtils.getNavigationLanguage(request);
 				
@@ -332,11 +332,11 @@ public class ActivityItem implements Comparable<ActivityItem>{
 	public void setAmounts(ActivityUtil.ActivityAmounts amounts) {
 		this.amounts = amounts;
 	}
-      public Long getPercent() {
+      public Float getPercent() {
         return percent;
     }
 
-    public void setPercent(Long percent) {
+    public void setPercent(Float percent) {
         this.percent = percent;
     }
